Fixed bug in b32 formatting.

+++ Tools/lsLog.h
@@ -42,7 +42,7 @@ s32 ls_vlogFormatPTR(char *dst, va_list *argList)
 
 s32 ls_vlogFormatB32(char *dst, va_list *argList)
 {
-    s64 intValue = va_arg(*argList, s64);
+    s32 intValue = va_arg(*argList, s32);
     
     if(intValue == FALSE) { ls_memcpy((void *)"false", dst, 5); return 5; }
     else                  { ls_memcpy((void *)"true", dst, 4);  return 4; }
